area converted to float (int64 is not serializable0 save the file AFTER converting to json. In case of errors won't destroy the file

+++ source/Blob.py
@@ -454,7 +454,7 @@ class Blob(object):
             self.perimeter += self.calculateContourPerimeter(self.contour)
 
     def calculateArea(self, mask):
-        self.area = mask.sum()
+        self.area = mask.sum().astype(float)
